Resolve #17, #18 -L6
